Two fifos now properly created.

+++ sauna.c
@@ -12,6 +12,9 @@
 
 #define FIFO_PERM 0700
 
+int generatorFD;
+int rejectedFD;
+
 int main(int argc, char* argv[]) {
     if (argc != 2){
 		fprintf(stderr, "Wrong usage: ./sauna <max capacity>\n");
@@ -24,9 +27,33 @@ int main(int argc, char* argv[]) {
 		fprintf(stderr, "Invalid capacity %d\n", maxCapacity);
 		exit(2);
 	}
-	
-    if(mkfifo("/tmp/rejeitados", FIFO_PERM) == -1)
-        perror("Error on creating FIFO rejeitados");    
 
+	if ((generatorFD = open("/tmp/entrada", O_RDONLY)) == -1){
+		perror("Fail on opening entrada for reading");
+		exit(3);
+	}
+	
+    if(mkfifo("/tmp/rejeitados", FIFO_PERM) == -1){
+        perror("Error on creating FIFO rejeitados");
+		exit(4);
+	}
+
+	if ((rejectedFD = open("/tmp/rejeitados", O_WRONLY | O_CREAT)) == -1){
+		perror("Error on opening rejeitados for writing");
+		exit(5);
+	}
+
+	pthread_t waitingLine;
+	pthread_create(&waitingLine, NULL, handleLine, );
+
+	//Missing stuff
+
+	ptread_join(waitingLine, NULL);
 	unlink("/tmp/rejeitados");
+}
+
+void * handleLine(void * arg){
+
+	queue * saunaLine = queue_constructor();
+
 }
